Split ParseUtils file
