+++ ConfigLoader/ConfigLoader.csproj
@@ -44,9 +44,11 @@
     <Compile Include="Extensions\StringSplitExtensions.cs" />
     <Compile Include="NameExports.cs" />
     <Compile Include="Utils\EnumUtils.cs" />
-    <Compile Include="Utils\ParseUtils.cs" />
+    <Compile Include="Utils\ParseOptions.cs" />
+    <Compile Include="Utils\ParseUtilsComplexTypes.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\IsExternalInit.cs" />
+    <Compile Include="Utils\ParseUtilsSimpleTypes.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="Assembly-CSharp">
